Use devenv.com instead of devenv.exe (#28)
For some reason on some Visual Studio installs devenv.exe just returns
immediately with an error code and runs the build in the background. It
also doesn't emit any output via stdout/stderr. According to [1]
devenv.com is the executable to use for build output via stdout/stderr
and it doesn't misbehave so switch to that.

[1]: https://docs.microsoft.com/en-us/visualstudio/ide/reference/devenv-command-line-switches?view=vs-2019